refactor: move injection, queue, supervision, and crash-insight primitives to relay-pty
Wave-2 extraction: relay-pty now carries the runtime-agnostic session
primitives alongside the PTY kernel, each decoupled from relay types via
a minimal generic or callback seam:

- queue: BoundedPriorityQueue is generic over the caller's priority
  scheme (PriorityScheme trait: LEVELS / FIRST_DROPPABLE / index;
  Prioritized gains an associated Priority type). The broker pins
  RelayPriority as a five-level scheme in priorities.rs.
- inject: Injector reports results through an observer callback
  (Arc<dyn Fn(&InjectResult)>) instead of EventEmitter, and
  deliver_with is generic over any request exposing an id via the new
  RequestId trait. InjectStatus/InjectResult move with their serde
  derives. (The broker module had been orphaned — never declared in
  lib.rs — so it gains compilation and tests by moving.)
- supervisor: the restart-policy state machine is generic over an
  opaque Clone respawn payload; the broker folds AgentSpec, parent,
  initial task, prompt options, and AgentResultMcpConfig into a
  SupervisedAgent payload in its supervisor.rs adapter.
- crash_insights: fully self-contained; moved verbatim.
- codex_session: takes the initialize-handshake client version as a
  parameter instead of reading the broker version helper.

Broker call sites adapt minimally: re-export shims in lib.rs, payload
field access in runtime/maintenance.rs, and the version argument at the
two worker.rs Codex spawn sites. specs/relay-pty-abstraction.md
documents the crate boundary, the target runtime-agnostic session
contract (terminal visualization as an optional capability, delivery
verification behind the receipt), and the remaining phases.
