+++ crates/relay-pty/Cargo.toml
@@ -11,11 +11,18 @@ path = "src/lib.rs"
 [dependencies]
 alacritty_terminal = "0.26"
 anyhow = "1.0"
+chrono = { version = "0.4", features = ["clock"] }
 parking_lot = "0.12"
 portable-pty = "0.8"
 regex = "1.11"
-tokio = { version = "1.44", features = ["sync", "time", "rt", "macros"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "2.0"
+tokio = { version = "1.44", features = ["sync", "time", "rt", "macros", "process", "io-util"] }
 tracing = "0.1"
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
+
+[dev-dependencies]
+tempfile = "3.19"
